feature: add possibility to remove tracks. Structure queries and mutations in different folders
